requirements.txt
discord.py-self==2.1.0
python-dotenv==1.2.1
pywebview==4.4.1
selenium==4.40.0
tk==0.1.0
